% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_POT_floods.R
\name{extract_pot_allstations}
\alias{extract_pot_allstations}
\title{Extracting POT-values for a set of stations}
\usage{

  extract_pot_allstations(amsfile = "inst/Example_data/Flooddata/amsvalues.txt",
  dailydata = "inst/Example_data/Dailydata", p_threshold = 0.98, TSEP = 6,
  pratio = 2/3, outfile = "inst/Example_data/Flooddata/potvalues.txt")
}
\arguments{
\item{amsfile}{file with annual maximum data. Only years with ams will be used}

\item{dailydata}{folder with daily data}

\item{p_threshold}{the thershold for selecting flood values given as emprical quantile of all data}

\item{pratio}{The minimum flow between two flood peaks should be less than pratio times the first flood peak.}

\item{outfile}{The file for storing ams values}

\item{TTR_3X}{The minimum time between two independent flood peaks}
}
\value{
data frame with reginenumber, main number, flood date, flod size, thershold. for all stations.
The dataframe is written to the specified outfile.
}
\description{
Use independence criterion based on Lang et al (1999)

See 'get_pot' for deitals on how independent POT values are extracted

Should first extract AMS values to get the years to be used for extracting POT data
}
\examples{
extract_pot_allstations(amsfile='inst/Example_data/Flooddata/amsvalues.txt',
dailydata="inst/Example_data/Dailydata",
p_threshold = 0.98, TSEP = 6,pratio= 2.0/3.0,
outfile="inst/Example_data/Flooddata/potvalues.txt")

}

