% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_recessiontimes.R
\name{extract_recessiontimes_allstations}
\alias{extract_recessiontimes_allstations}
\title{Calculate recession times for a set of stations}
\usage{
extract_recessiontimes_allstations(fraction = 0.995,
  stations_periods_file = "inst/Example_data/Flooddata/Table_stations_periods.csv",
  dailydata = "inst/Example_data/Dailydata",
  outfile = "inst/Example_data/Flooddata/recessiontimes.txt")
}
\arguments{
\item{fraction}{quantile of the pdf of recession coefficients lamda used in calculation. A high value is recommended}

\item{stations_periods_file}{a file with list of stations and data periods to be used for estimationg recession}

\item{dailydata}{folder with daildata}

\item{outfile}{file for storing results}
}
\value{
a data object with recession times in days for all catchments
}
\description{
See 'get_recession_time' for details on how recession times are extracted
}
\examples{
extract_recessiontimes_allstations(fraction=0.995,
"inst/Example_data/Flooddata/Table_stations_periods.csv",
"inst/Dailydata", "inst/Example_data/Flooddata/recessiontimes.txt")

}

