% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ams_functions_v2.R
\name{get_amsdata}
\alias{get_amsdata}
\title{Extract annual maxima from daily time series for specified years rpovided the year has more than 363 observations.}
\usage{
get_amsdata(stationnumber = 200011,
  path_dd = "../inst/Excample_data/Dailydata",
  path_sd = "../inst/Excample_data/Subdaily", active_station = 0,
  d_first = 1880, d_last = 2015, d_exclude = NA, s_first = 1880,
  s_last = 2015, s_exclude = NA)
}
\arguments{
\item{path_dd}{path to folder with daily data}

\item{path_sd}{path to folder with subdaily data}

\item{d_first}{first year to be used for daily data}

\item{d_last}{last year to be used for daily data}

\item{d_exclude}{years to be excluded for daily data}

\item{s_first}{first year to use for subdaily data}

\item{s_last}{last year to use for subdaily data}

\item{s_exclude}{years to exclude for subdaily data}
}
\value{
dataframe with station numbers, date an values for for daily and subdaily annual maxima
If they are from teh same event, they are written on the same line
if they are from different events, they are written on different lines
}
\description{
Extract annual maxima from daily time series for specified years rpovided the year has more than 363 observations.
}
\examples{

}

