% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FGP.R
\name{get_fgp}
\alias{get_fgp}
\title{Extracting flood generating process as the fraction of flood runoff casued by rain (0-1)
for a list of floods specified by dates for one station}
\usage{
get_fgp(cfloods, crain, csnow, cdates, ctime = 2, rtime = 6)
}
\arguments{
\item{cfloods, }{data with floods. amsfile file with the flood data.}

\item{crain}{raindata for one catchmnents}

\item{csnow}{snowmeltdata for the catchment}

\item{rtime}{recession time for one station}

\item{Dates}{for the Metdata}

\item{Ctime}{concentration time of the catchments.Is 2 days for all catchments}
}
\value{
vector with fgp for ecah of the stations.
}
\description{
Extracting flood generating process as the fraction of flood runoff casued by rain (0-1)
for a list of floods specified by dates for one station
}
\examples{
 sfgp<-get_fgp(cfloods=floods[snames_floods==mystations[i],],crain=raindata[,loc_rain],
csnow=snowdata[,loc_rain],cdates=metdates,ctime=2,rtime=recession[i,4]),ncol=1)


}

