% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FGP.R
\name{get_fgp_allstations}
\alias{get_fgp_allstations}
\title{Extracting flood generating process for a set of stations}
\usage{
get_fgp_allstations(amsfile = "inst/Example_data/Flooddata/amsvalues.txt",
  rainfile = "inst/Example_data/klimadata/aveR.txt",
  snowfile = "inst/Example_data/klimadata/aveS.txt",
  recessionfile = "inst/Example_data/flooddata/recessiontimes.txt",
  outfile = "inst/Example_data/Flooddata/ams_and_fgp.txt")
}
\arguments{
\item{amsfile}{file with the flood data. Created by the function 'extract_ams_allstations'}

\item{rainfile}{file with the rain data. Created by the function 'get_metdataforfloods'}

\item{snowfile}{file with the snow melt data. Created by the function 'get_metdataforfloods'}

\item{recessionfile}{file with recession times. Created by the function 'extract_recessiontimes_allstations'}

\item{outfile}{file for storing the fgp. Similar to the input flood file, but with a colomn of fgp added.}
}
\value{
dataframe with floods where fgp is included
}
\description{
Flood generating process is calculated as as the fraction of flood runoff casued by rain (0-1)
for a set of stations and identified flood events. time series of catchment average rain and snow melt is
extracted from the SeNorge model and is performed using the function 'extract_ams_allstations'. The recession times
are calculated using 'extract_recessiontimes_allstations'
 See "get_fgp" for details
}
\examples{
get_fgp_allstations(amsfile='inst/Example_data/Flooddata/amsvalues.txt',rainfile='inst/Example_data/klimadata/aveR.txt',
snowfile='inst/Example_data/klimadata/aveS.txt',recessionfile='inst/Example_data/flooddata/recessiontimes.txt',
outfile='inst/Example_data/Flooddata/ams_and_fgp.txt')

}

