% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMetValuesForFloods.R
\name{get_metdataforfloods}
\alias{get_metdataforfloods}
\title{Title Extract daily values for precipitation, rain, snowmelt, temperature and runoff from
the SeNorge data for specified catchments.}
\usage{
get_metdataforfloods(gridid = NA, first_day = as.Date("1961/1/1"),
  last_day = as.Date("1990/12/31"),
  station_file = "../Data/Flooddata/Table_stations_periods.csv",
  snr_translation = "../Data/Excample_data/CatchmentCharacteristics/Feltnr_flomkart_til_feltnr_GIS.txt",
  metfolder = "U:/metdata/", snowfolder = "U:/snowsim/",
  hbvfolder = "Z:/gwbsim/", outfolder = "../Data/Complete_data/Flooddata/")
}
\arguments{
\item{gridid}{List of SeNorge grids for each catchment. Created by the function gridcell_list}

\item{first_day}{The first day of the time sequence}

\item{last_day}{The last day of the time sequence}

\item{station_file}{List of stations if only a subset from the gridid-list is needed}

\item{snr_translation}{Translation to GIS-station numbers if necesessary}

\item{metfolder}{folder where SeNorge Met-grids are stored}

\item{snowfolder}{folder where the SeNorge snow grids are stored}

\item{hbvfolder}{folder where the SeNorge water balance modellin grids are stored}

\item{outfolder}{folder for witing the results}
}
\value{
list with tables of daily and average monthly values
}
\description{
Title Extract daily values for precipitation, rain, snowmelt, temperature and runoff from
the SeNorge data for specified catchments.
}
\examples{
get_metdataforfloods(gridid=grid_id_all_catchments,first_day=as.Date("1961/1/1"),last_day=as.Date("1961/12/31"),
station_file="../inst/Excample_data/Flooddata/Table_stations_periods.csv",
snr_translation="../inst/Excample_data/CatchmentCharacteristics/Feltnr_flomkart_til_feltnr_GIS.txt",
metfolder="U:/metdata/",snowfolder="U:/snowsim/",hbvfolder="Z:/gwbsim/",outfolder="../inst/Complete_data/Flooddata/")
}

