% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_POT_floods.R
\name{get_pot}
\alias{get_pot}
\title{Extract independent POT values for one station}
\usage{
get_pot(snumb = 200011, f_years = NA,
  path_dd = "inst/Example_data/Dailydata", p_threshold = 0.98, TSEP = 6,
  pratio = 2/3)
}
\arguments{
\item{snumb}{Station number accordin to NVE: rrmmmmm where r is reine  number and m is main number}

\item{f_years}{Vectors of years that should be used. If NA, all years in the data file will be used}

\item{path_dd}{Folder with daily data}

\item{p_threshold}{The thershold for selecting flood values given as a precentile}

\item{pratio}{The minimum flow between two flood peaks should be less than pratio times the first flood peak.}

\item{TTR_3X}{The minimum time between two independent flood peaks}
}
\value{
data frame with reginenumber, main number, flood date, flod size, thershold for the station specified by snumb.
}
\description{
Use independence criterion based on Lang et al (1999)

The following calculation steps are used:

1: Find the threshold T based on the pecentile P from the observed daily data

2: Find all values abouve the thershold and identify in which direction the threshold is crossed

3: Find maximums of all clusters above thershold. A cluster is all values between an upward and a downward crossing.

4: Find the number of days between each maximum from 3

5: If data are separated by less than TTR*3, select the maximum. TTR is 'time to rize'. See Lang et al (1999)

6: Find the minimum flow between two flood events

7: if the minimum flow is higher than 2/3 of the first flood, then select the two

10.06.2016, LESC (lena schlichting)
}
\examples{
extract_pot_allstations(amsfile='inst/Example_data/Flooddata/amsvalues.txt',
dailydata="inst/Example_data/Dailydata",
p_threshold = 0.98, TTR_3x = 6,pratio= 2.0/3.0,
outfile="inst/Example_data/Flooddata/potvalues.txt")
}

