% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_recessiontimes.R
\name{get_recession_time}
\alias{get_recession_time}
\title{Title}
\usage{
get_recession_time(fraction = 0.995, stationnumber = 200011,
  path_dd = "../Dogndata", active_station = 0, d_first = 1880,
  d_last = 2015, d_exclude)
}
\arguments{
\item{fraction}{quantile of the pdf of recession coefficients lamda used in calculation. A high value is recommended}

\item{stationnumber}{station number given as}

\item{path_dd}{path for daily data}

\item{active_station}{if active station is 1, ignore d_last and use data until end of data period}

\item{d_first}{first year for extracting recession data}

\item{d_last}{last year for extracting recession data}

\item{d_exclude}{years to be excluded}
}
\value{
recession time in days for the sepcified catchment
}
\description{
Title
}

