% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ams_functions_v2.R
\name{get_subdaily_period}
\alias{get_subdaily_period}
\title{Extract period with subdaily data from a data with unregular times}
\usage{
get_subdaily_period(s_dat)
}
\arguments{
\item{s_dat}{the data frame with the dates}
}
\value{
a dataframe where only periods with subdaily data is included
}
\description{
Extract period with subdaily data from a data with unregular times
}

