% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeNorge_gridID_for_catchments.R
\name{gridcell_list}
\alias{gridcell_list}
\title{Extracting SeNorge grid-ids for catchments}
\usage{
gridcell_list(c_ids = c("2.11.0", "12.13.0"),
  c_shape = "Hydrologi_TotalNedborfeltMalestasjon.shp", c_layer = NA,
  snr_translation = "inst/Example_data/CatchmentCharacteristics/Feltnr_flomkart_til_feltnr_GIS.txt",
  llcenter = c(-74500, 6450500), gridsize = c(1000, 1000),
  griddim = c(1195, 1550), myprojection = NA,
  outfile = "inst/Example_data/GISData/CID.txt")
}
\arguments{
\item{c_ids}{The lis of catchmen ID's that is specified as "RegineNumber.MainNumber.0"}

\item{c_shape}{The shape file with the catchment boundaries for the gauging stations}

\item{c_layer}{The layer of the shape file that is to be extracted. Normally the same name as the shape-file.}

\item{llcenter}{Te lower left cell center of the SeNorge grid system (UTM33N)}

\item{gridsize}{Gridsize in meters}

\item{griddim}{Number of gridcells in x and y directions}

\item{myprojection}{Projection might be specified. If not use the proj4-string from c_shape.}

\item{outfile}{File where the GridCell ids for each catchment is stored}
}
\value{
a list of grid-ids, where each element of the list represent a catchment. The names of the grid elemnts are copied from c_ids"RegineNumber.MainNumber.0"
}
\description{
The SeNorge grid ids starts at 0 or 1 (depending on coding language) in upper left corner,
and increase line by line. The coordinates are given in UTM33N and the center of the lower left cell
has coordinates c(-74500, 6450500). Grid size is 1000 meters. The dimensjon of the grid is 1195,1550
Shape file with catchment boundaries for NVE gauging stations might be downloaded from
http://nedlasting.nve.no/gis/, and you should select HYDROLOGISKE DATA->Totalnedbørfelt til målestasjon
}
\examples{
grid_id_Narsjo<-gridcell_list("2.11.0",
"//nve/fil/h/HM/Interne Prosjekter/Flomkart/Data/GISData/Hydrologi_TotalNedborfeltMalestasjon.shp",
c_layer="Hydrologi_TotalNedborfeltMalestasjon",
snr_translation="inst/Example_data/CatchmentCharacteristics/Feltnr_flomkart_til_feltnr_GIS.txt"
outfile='inst/Example_data/GISData/CID.txt')

grid_id_all_catchments<-gridcell_list(NA,
"//nve/fil/h/HM/Interne Prosjekter/Flomkart/Data/GISData/Hydrologi_TotalNedborfeltMalestasjon.shp",
c_layer="Hydrologi_TotalNedborfeltMalestasjon",
snr_translation="inst/Excample_data/CatchmentCharacteristics/Feltnr_flomkart_til_feltnr_GIS.txt"
outfile='inst/Example_data/GISData/CID.txt')
}

