% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ams_functions_v2.R
\name{insertRow}
\alias{insertRow}
\title{Inserts an extra line in the data frame with flood data. This is necessary if for one year the
date of the annual maximum daily flood is more than one day different than that for the subdaily annual maximum}
\usage{
insertRow(existingDF, r, dailydat, knekkdat)
}
\arguments{
\item{existingDF}{the existing data frame}

\item{r}{the index where the new row should be incerted}

\item{dailydat}{data frame with daily data}

\item{knekkdat}{data frame with subdaily data}
}
\value{
data frame with one more line than teh input
}
\description{
Inserts an extra line in the data frame with flood data. This is necessary if for one year the
date of the annual maximum daily flood is more than one day different than that for the subdaily annual maximum
}
\examples{
insertRow(existingDF,r,dailydat,knekkdat)
}

