% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_recessiontimes.R
\name{recession_times}
\alias{recession_times}
\title{Calculates recession times}
\usage{
recession_times()
}
\description{
## This script calculates average recession times in days, based on method described in Skaugen & Onof, 2014.
}

