% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_recessiontimes.R
\name{sum_nona}
\alias{sum_nona}
\title{Function for calculating number of values that are not NA in a vector}
\usage{
sum_nona(xx)
}
\arguments{
\item{xx}{}
}
\value{
integer
}
\description{
Function for calculating number of values that are not NA in a vector
}

