% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_update_AMStable.R
\name{UPDATE_AMS_TABLE_TODO}
\alias{UPDATE_AMS_TABLE_TODO}
\title{This script updates the AMS table with flood peak for daily data and hourly data, as well as corresponding data @flood if knekk and daily arent same event}
\usage{
UPDATE_AMS_TABLE_TODO()
}
\description{
This script updates the AMS table with flood peak for daily data and hourly data, as well as corresponding data @flood if knekk and daily arent same event
}

