% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ams_functions_v2.R
\name{extract_ams_allstations}
\alias{extract_ams_allstations}
\title{Extracting ams-values for a list of stations}
\usage{

  extract_ams_allstations(stations_periods_file = "../Data/Flooddata/Table_stations_periods.csv",
  dailydata = "../Data/Dailydata", subdailydata = "../Data/Subdaily",
  outfile = "../Data/Flooddata/amsvalues.txt")
}
\arguments{
\item{stations_periods_file}{file with station numbers and perios for data}

\item{dailydata}{folder with daily data}

\item{subdailydata}{folder with subdaily data}

\item{outfile}{file for stopring the ams values}
}
\description{
Extracting ams-values for a list of stations
}

