% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_recessiontimes.R
\name{extract_recessiontimes_allstations}
\alias{extract_recessiontimes_allstations}
\title{Wrapping the 'get_recession_time' function to get recession times for a list of stations.
Calculates average recession times in days, based on method described in Skaugen & Onof, 2014.
Skaugen, T. and Onof, C. (2014): A rainfall-runoff model parameterized from GIS and runoff data. Hydrol. Process., 28, 4529-4542.}
\usage{
extract_recessiontimes_allstations(fraction = 0.995,
  stations_periods_file = "../Data/Flooddata/Table_stations_periods.csv",
  dailydata = "../Data/Dailydata",
  outfile = "../Data/Flooddata/recessiontimes.txt")
}
\arguments{
\item{fraction}{quantile of the pdf of recession coefficients lamda used in calculation. A high value is recommended}

\item{stations_periods_file}{a file with list of stations and data periods to be used for estimationg recession}

\item{dailydata}{folder with daildata}

\item{outfile}{file for storing results}
}
\value{
a data object with recession times in days for all catchments
}
\description{
Wrapping the 'get_recession_time' function to get recession times for a list of stations.
Calculates average recession times in days, based on method described in Skaugen & Onof, 2014.
Skaugen, T. and Onof, C. (2014): A rainfall-runoff model parameterized from GIS and runoff data. Hydrol. Process., 28, 4529-4542.
}
\examples{
extract_recessiontimes_allstations(fraction=0.995,
"../inst/Example_data/Flooddata/Table_stations_periods.csv",
"../inst/Dailydata", "../inst/Example_data/Flooddata/recessiontimes.txt")
}

