% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FGP.R
\name{get_fgp_allstations}
\alias{get_fgp_allstations}
\title{Extracting flood generating process as the fraction of flood runoff casued by rain (0-1)
for a list of floods specified by station numbers and dates}
\usage{
get_fgp_allstations(amsfile = "inst/Example_data/Flooddata/amsvalues.txt",
  rainfile = "inst/Example_data/klimadata/aveR.txt",
  snowfile = "inst/Example_data/klimadata/aveS.txt",
  recessionfile = "inst/Example_data/flooddata/recessiontimes.txt",
  outfile = "inst/Example_data/Flooddata/ams_and_fgp.txt")
}
\arguments{
\item{amsfile}{file with the flood data. Created by the function extract_ams_allstations}

\item{rainfile}{file with the rain data. Created by the function get_metdataforfloods}

\item{snowfile}{file with the snow melt data. Created by the function get_metdataforfloods}

\item{outfile}{file for storing the fgp. Similar to the input flood file, but with a colomn of fgp added.}
}
\value{
dataframe with floods where fgp is included
}
\description{
Extracting flood generating process as the fraction of flood runoff casued by rain (0-1)
for a list of floods specified by station numbers and dates
}
\examples{
get_fgp_allstations(amsfile='inst/Example_data/Flooddata/amsvalues.txt',rainfile='inst/Example_data/klimadata/aveR.txt',
snowfile='inst/Example_data/klimadata/aveS.txt',recessionfile='inst/Example_data/flooddata/recessiontimes.txt',
outfile='inst/Example_data/Flooddata/ams_and_fgp.txt')

}

