% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ams_functions_v2.R
\name{get_subdaily_period}
\alias{get_subdaily_period}
\title{Extract period with subdaily data from a data with unregular times. Looks for the first four
days in a row with subdaily time resolution. Assumes the reimaining period is at subdaily resoultion}
\usage{
get_subdaily_period(s_dat)
}
\arguments{
\item{s_dat}{the data frame with the dates}
}
\value{
a dataframe where only periods with subdaily data is included
}
\description{
Extract period with subdaily data from a data with unregular times. Looks for the first four
days in a row with subdaily time resolution. Assumes the reimaining period is at subdaily resoultion
}
\examples{
get_subdaily_period(s_dat)
}

